[FIX/#62] 자동 로그인 로직 MainActivity로 이동
