+++ app/src/main/java/com/growthook/aos/presentation/onboarding/LoginActivity.kt
@@ -23,7 +23,6 @@ class LoginActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = ActivityLoginBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        isAlreadyLogin()
         startKakaoLogin()
         isKakaoLogin()
     }
@@ -38,19 +37,6 @@ class LoginActivity : AppCompatActivity() {
         viewModel.isLoginSuccess.observe(this) { isSuccess ->
             if (isSuccess) {
                 val intent = Intent(this, MainActivity::class.java)
-                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
-                startActivity(intent)
-                finish()
-            }
-        }
-    }
-
-    private fun isAlreadyLogin() {
-        viewModel.checkIsAlreadyLogin()
-        viewModel.isAlreadyLogin.observe(this) { isAlreadyLogin ->
-            if (isAlreadyLogin) {
-                val intent =
-                    Intent(this@LoginActivity, MainActivity::class.java)
                 intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
                 startActivity(intent)
                 finish()
